Revert "Feature/basic auth (#20)"
In retrospect, this added complexity that we don't use at comma.

Basic auth is sadly in between what would be useful, like signed URLs with a permission model, and nothing. The issue is that if it's compromised once, the whole server is compromised, including for writes. I should have considered this more before agreeing to add it.

In the future, upstream minikeyvalue is stated to only be designed for firewalled networks.

Since it's not being used in production, I've decided to revert this commit.

This reverts commit cbfba5dd106b328f6a44a633f0ef96b1122b5227.